Provide module Ki.Typed to prevent escape
RankNTypes encode the notion of an inviolable /scope/ at the type level.
As `runST` famously demonstrates, this can be utilised to disallow
value-level scope violations at compile time.

`Scope` and `Thread` are invalid outside their parent scope, hence a new
interface is provided to enforce a scoping invariant. However, in order
for the internals to leverage that invariant, the original interface
must go.